fixed swerve CAN ids, replaced sensor talons with CANcoders
the swerve CAN ids are now standardized

the sensors on the swerves are now read through the CANcoder system (direct through CAN). we tested the difference in lagginess between the falcon encoder and the CANcoder values, and they were very similar, but CANcoder seemed a bit more stable at 20ms read time than the falcon. Interestingly, the falcons seemed to read more consistently and a bit more frequently than the CANcoders at 1ms read time, but this would never be used