removed test compressor in p-tasks, custom flywheel control, cowl sensor
the old code for testing the compressor was still there, so i removed it

added a way to control the power of the flywheel with a numeric control from the shooter_test front panel, and button to enable and disable whether we use %vbus or the new addition in the test mode

added a conversion of the cowl sensor to the angle of the cowl, just using the values we measured physically. didn't really need to do it now, but it's there. i was gonna use it for a limit, but we don't really need it right now