added fly & magazine speed control code, begin flywheel PID, fly sensor
flywheel speed control works, but there isn't an encoder on the magazine. we need one in order to speed control it. added speed control testing options for magazine in feeder_determiner, and the same for flywheel in shooter_determiner. probably want to add that for the cowl as well. 

the flywheel PID constants are set in begin

added flywheel velocity reads on left and right, and averaged them, as well as the conversion between raw sensor ticks per 100 ms to RPM and back for direct speed control to the talons