Commit WITH calibration code
I am committing with the calibrations code first to have a basis to add in the sensors. This is pre-sensors

For calibration code:
I've added global variables for every subsystem on the dashboard. -Drive has some problems, I may rearrange how its done. Begin and period tasks are not done yet. They have a basis, but have some errors and unfinished parts.