changed magazine shoot mode to %Vbus for now, other changes
we struggled to get speed control to work, so for now we're just gonna use %Vbus and see if its bad or not.

also added new cowl limits, the cowl has to be pushed all the way down before the robot is booted. we'll probably have to add a "zero all" button and directions of how to zero everything for the drivers, especially for the climbers. 

there are probably other changes, but i don't know them.