rudimentary climber sequencer, gross PID tuning control state
almost works, the end of the sequence was shortened because the initial pull of the outer arms on the 3rd bar seemed to pull the inner arms off relatively consistently

the 0th, 1st, and 2nd steps of the main sequence seem to not be consistently getting the robot onto the bar, partly because the elevator does not hold at its position with the brake engaged, as well as the outer arm pivot PID doesn't get exactly to 0 because of the slack of the gears. we'll try to tune a better one tomorrow. mentors say that if the input to the PID of the outer arm pivot was an IMU, we'd be much more accurate on our arm position, and the inner arms would be at a more consistent spot in space. we'll see. also, the outer arm hooks keep ramming into the 2nd bar very violently, which I'm not sure if it ruins the sensor value or not. 

the tune PID control state is terrible, awful, no good, very bad, but i don't have time to make it better. it's usable at least.