changed drive joystick tuning a bit, prevent tipping, path planning good
the swerve control were way too sensitive, so we reduced the max values for the rotational throttle, and brought the exponential on the strafe throttles from 3 to 1, because we just weren't moving when the joystick was near zero.

we increased the drive motor ramp rate slightly so that now tipping flat out is not a big issue. however, whenever we change our target angle very quickly, the robot will still tilt because the robot can't change directions that fast. not sure how to confront that, might ask for help

tested out the path planning, i screwed up autonomous load with reading the string in DrivePath, now it only reads the string when drivepath is passed in, and reads parameters normally otherwise