fixed step 1 in climb sequence
idk why i put in an IMU deg goal for leveling the pivot, the robot should be perpendicular to the outer arms for the inner arms, not to the ground. also, the PID loop was not enabled, and the PID values wouldn't have worked regardless because I didn't retune them. not sure when i did that, but it was pretty dumb. step 1 pivot PID is enabled now, and uses the outer arm encoder for the PID, works much more consistently