fixed cowl issue, modified IMU-based elevator steps
the cowl was hitting a phantom limit switch that it "read" from the gadgeteer, so i disabled the limit switches in begin.

i made it so that the elevator won't retract back to the short length after extending to full length when going towards the next bar, and I made it so that the elevator won't extend after fully doing the pull-up back to the short length. also, i made it so that the steps cannot end when the short distance is being used, since there was a weird ending step at one point