added cowl movement in turret control command, reverted 5-ball changes
i have an idea for a different 5-ball where we pick up 2 balls first, shoot, pick up 1, shoot, then go to HPP, but i need to commit before i make the new one.

the modified 5-ball that i was going to create but reverted had the 3-ball not spin at the end to face the target, but apparently that spin is pretty important to actually get the balls into the intake, because the balls kept lining up side-by-side in the funnel if i didn't do that turn.